refactor(desktop): remove crate-wide clippy allow block and fix its 46 warnings (R-44) [skip ci]
main.rs:2 disabled 15 clippy lints for the whole glucose-desktop crate, which
made `cargo clippy --all-targets -- -D warnings` prove nothing: it validated a
crate whose alarms had been switched off first. The first lint of the list,
too_many_arguments, would have flagged render_docks() — 11 parameters, 5 of them
consecutive f32 — before it was called with mouse_x (170) in the `scale` slot,
rasterizing fonts at 2040 px and freezing the window at 10 697 ms per frame.

Five of the fifteen lints no longer suppressed anything and were dead weight:
manual_is_multiple_of, manual_range_contains, collapsible_if,
collapsible_else_if, single_match.

The 46 warnings the block was hiding are fixed, none of them by an #[allow]:

- too_many_arguments (17): related parameters grouped into named types in the
  new params.rs module — Pointer, ScreenFrame, ScaledRect, ButtonState,
  SceneOverlay, ViewPass — plus TextStyle in typography.rs. Each is built with
  explicit named fields at the call site, so passing a position where a scale is
  expected is now a compile error.
- unnecessary_cast (12), field_reassign_with_default (7), type_complexity (2),
  collapsible_match (2), manual_strip (2), chunks_exact_to_as_chunks (1),
  derivable_impls (1), new_without_default (1), unwrap_or_default (1).

No behaviour change: 261 tests green, frames measured at ~16 ms under
GLUCOSE_PERF=2. The three oversized files shrank: dock.rs 2067 -> 2016,
ui.rs 1356 -> 1348, renderer.rs 1179 -> 1165.

+++ crates/glucose-desktop/src/icons.rs
@@ -253,7 +253,7 @@ pub fn draw_icon_scaled(
                     (color.red() * 255.0) as u8,
                     (color.green() * 255.0) as u8,
                     (color.blue() * 255.0) as u8,
-                    ((color.alpha() * 255.0) as f32 * 0.45) as u8,
+                    (color.alpha() * 255.0 * 0.45) as u8,
                 );
                 faint_paint.set_color(col);
                 faint_paint.anti_alias = true;
